+++ lib/src/screens/qr_scan_screen.dart
@@ -59,7 +59,7 @@ class _QrScanScreenState extends State<QrScanScreen> {
               child: const Padding(
                 padding: EdgeInsets.all(14),
                 child: Text(
-                  'Наведи камеру на QR с подпиской, VLESS Reality или Hysteria ссылкой.',
+                  'Наведи камеру на QR с подпиской, VLESS Reality, NaiveProxy или Hysteria ссылкой.',
                   textAlign: TextAlign.center,
                 ),
               ),